LayoutTests/imported/w3c/web-platform-tests/workers/shared-worker-name-via-options-expected.txt

CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'worker1.port.onmessage = receiveMessage')

FAIL Test the name property of shared workers mixing constructor options and constructor strings TypeError: null is not an object (evaluating 'worker1.port.onmessage = receiveMessage')
